ops/synthetic_probes/requirements.txt
httpx==0.27.2
PyYAML==6.0.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
